+++ CHANGELOG.md
@@ -6,7 +6,7 @@ ProofNetworkJdf Changelog
  * --
 
 #### Bug Fixing
- * --
+ * XJDF: Fix missing namespaces (for updateFrom)
 
 ## 0.19.8.7
 #### Features